+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - Use fixed codegen version.
   - Support for confugurable from env API url.
   - Support for set domain ignore and max emojii.
+  - Default ignore domains.
 
 2024-01-10 v0.1.6
   - Temporary fix domain.